requirements.txt
livekit>=1.5.0
livekit-agents>=1.4.0
livekit-plugins-google>=1.3.11
livekit-plugins-deepgram>=1.3.11
livekit-plugins-silero>=1.3.11
python-dotenv>=1.0.1